created notebook for ranking
